requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
